+++ app/Http/Controllers/Administrator/CategoryController.php
@@ -8,7 +8,21 @@ use Illuminate\Http\Request;
 
 class CategoryController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            if (!auth('admin')->check()) {
+                return redirect()->route('login');
+            }
+            return $next($request);
+        });
+    }
+
     public function index(){
+        if (!auth('admin')->user()->hasPermissionTo('Categories Access')) {
+            abort(403, 'Unauthorized action.');
+        }
+        
         $data = array(
             'title'         => 'Category',
             'categories'    => Category::latest()->get(),
@@ -17,6 +31,15 @@ class CategoryController extends Controller
     }
 
     public function store(Request $req){
+        if (isset($req->category_id) && !empty($req->category_id)) {
+            if (!auth('admin')->user()->hasPermissionTo('Categories Edit')) {
+                abort(403, 'Unauthorized action.');
+            }
+        } else {
+            if (!auth('admin')->user()->hasPermissionTo('Categories Create')) {
+                abort(403, 'Unauthorized action.');
+            }
+        }
         
         $req->validate([
                 'name'          => ['required', 'max:255'],
@@ -41,6 +64,10 @@ class CategoryController extends Controller
     }
 
     public function edit($id){
+        if (!auth('admin')->user()->hasPermissionTo('Categories Edit')) {
+            abort(403, 'Unauthorized action.');
+        }
+        
         $data = array(
             'title'         => 'Category',
             'categories'    => Category::latest()->get(),
@@ -51,6 +78,10 @@ class CategoryController extends Controller
     }
 
     public function delete($id){
+        if (!auth('admin')->user()->hasPermissionTo('Categories Delete')) {
+            abort(403, 'Unauthorized action.');
+        }
+        
         Category::destroy(hashids_decode($id));
         return response()->json([
             'success'   => 'Category deleted successfully',
