+++ app/Http/Controllers/Administrator/PermissionController.php
@@ -10,25 +10,39 @@ use Illuminate\Support\Facades\Validator;
 
 class PermissionController extends Controller
 {
-    // public function __construct()
-    // {
-    //     $this->middleware('is_admin');
-        
-    // }
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            if (!auth('admin')->check()) {
+                return redirect()->route('login');
+            }
+            return $next($request);
+        });
+    }
 
     public function index(Request $request)
     {
+        if (!auth('admin')->user()->hasPermissionTo('Permissions Access')) {
+            abort(403, 'Unauthorized action.');
+        }
+        
         $data = array(
             'title' => 'All Permissions',
             'permissions' => Permission::orderby('id', 'desc')->get(),
         );
         return view('admin.permissions.all_permissions')->with($data);
-        // return view('admin.ranges.add_range')->with($data);
-
     }
 
     public function save(Request $request, Slug $slug)
     {   
+        if (!$request->permission_id && !auth('admin')->user()->hasPermissionTo('Permissions Create')) {
+            abort(403, 'Unauthorized action.');
+        }
+        
+        if ($request->permission_id && !auth('admin')->user()->hasPermissionTo('Permissions Edit')) {
+            abort(403, 'Unauthorized action.');
+        }
+        
         $rules = [
             'name' => ['required', 'string', 'max:80'],
         ];
@@ -64,6 +78,10 @@ class PermissionController extends Controller
 
     public function delete(Request $request)
     {
+        if (!auth('admin')->user()->hasPermissionTo('Permissions Delete')) {
+            abort(403, 'Unauthorized action.');
+        }
+        
         $permission = Permission::hashidOrFail($request->permission_id);
         $permission->delete();
         return response()->json([
